tests: use ZooKeeper from upstream for tests
This removes the dependency on docker that is problematic when already
inside a container and the duplication of ZooKeeper fixtures.
