+++ .github/workflows/tests.yml
@@ -18,7 +18,7 @@ jobs:
         python-version: ${{ matrix.python-version }}
 
     - name: Install dependencies
-      run: make build-dep-ubuntu
+      run: make build-dep-ubuntu test-dep-ubuntu
 
     - name: Install requirements
       run: python3 -m pip install -r requirements.txt
